third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-link-element/link-style-data-uri-quirks-expected.txt

This is a testharness.js-based test.
[FAIL] data: URI stylesheet with text/plain is blocked in standards mode
  assert_equals: expected "error" but got "load"
Harness: the test ran to completion.
